requirements.txt
pytest==7.4.3
pytest-playwright==0.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-reportportal==5.6.0
playwright==1.40.0
python-dotenv==1.0.0
